Create Average Distance.py
Chefland is a grid with N rows and M columns. Each cell of this grid is either empty or contains a house. The distance between a pair of houses is the Manhattan distance between the cells containing them.
For each d between 1 and N+M−2 inclusive, Chef wants to calculate the number of unordered pairs of distinct houses with distance equal to d.
https://www.codechef.com/SNCK1A19/problems/AVGMAT